TSDK-820 Continue refactoring
