+++ topl-btc-bridge-consensus/src/main/scala/co/topl/bridge/ConsensusParamsDescriptor.scala
@@ -6,7 +6,7 @@ import org.bitcoins.core.currency.SatoshisLong
 import co.topl.brambl.models.SeriesId
 import co.topl.brambl.models.GroupId
 
-trait BridgeParamsDescriptor {
+trait ConsensusParamsDescriptor {
 
   import ParamParser._
