+++ topl-btc-bridge-public-api/src/main/scala/co/topl/bridge/Main.scala
@@ -1,0 +1,37 @@
+package co.topl.bridge
+
+import cats.effect.ExitCode
+import cats.effect.IO
+import cats.effect.IOApp
+import scopt.OParser
+
+sealed trait PeginSessionState
+
+case object PeginSessionState {
+  case object PeginSessionStateWaitingForBTC extends PeginSessionState
+  case object PeginSessionStateMintingTBTC extends PeginSessionState
+  case object PeginSessionWaitingForRedemption extends PeginSessionState
+  case object PeginSessionWaitingForClaim extends PeginSessionState
+  case object PeginSessionMintingTBTCConfirmation extends PeginSessionState
+  case object PeginSessionWaitingForEscrowBTCConfirmation
+      extends PeginSessionState
+  case object PeginSessionWaitingForClaimBTCConfirmation
+      extends PeginSessionState
+}
+
+object Main extends IOApp with PublicApiParamsDescriptor {
+
+  override def run(args: List[String]): IO[ExitCode] = {
+    OParser.parse(
+      parser,
+      args,
+      ToplBTCBridgePublicApiParamConfig()
+    ) match {
+      case Some(_) =>
+        IO(ExitCode.Success)
+      case None =>
+        IO.consoleForIO.errorln("Invalid arguments") *>
+          IO(ExitCode.Error)
+    }
+  }
+}
